fix(tests): the fleet-store guard sat one directory too deep
MEASURED, and it is a recurrence of the incident the guard's own docstring
describes. `_no_accidental_fleet_store_writes` lived in
`tests/scitex_agent_container/conftest.py`. `tests/smoke/`, `tests/develop/`,
`tests/integration/` and `tests/examples/` are its SIBLINGS, not its children,
so none of them was ever covered. When this branch moved `acl_deny_notify` to
PostgreSQL, the smoke suite began writing its fixtures into the live per-host
store — exactly what the lifecycle tests did under #1154, one directory over.

WHAT WAS IN PRODUCTION, on scitex-compute-04:

    postgresql://scitex_cards@127.0.0.1:55432/scitex   acl_deny_notify_log
      sender='alpha'  target='gamma'
      sender='alpha'  target='beta'

Both written by a bare `pytest tests/smoke/...` run. They have been hidden
with the store's own `hide` verb (actor recorded, oplog intact), and a re-read
on a fresh handle reports 0 visible rows.

THE SECOND-ORDER FAILURE IS THE INTERESTING ONE. The notification is
rate-limited per (sender, target) with a 30-minute cool-down. Under SQLite that
state was a tmp file per test. Under a shared PostgreSQL it persists, so the
FIRST run of `test_listen_denied_send_persists_two_channel_events_rows` passed
and every run within the next 30 minutes saw the notification suppressed and
asserted `1 == 2`. A test that passes once and then fails for half an hour
reads as flaky infrastructure; it was the migration.

WHY THE WHOLE SUITE LOOKED FINE. Collecting the `scitex_agent_container`
package imports its conftest, and the module-level assignment there is
process-wide, so a full-suite run was protected by accident of import order.
Running `tests/smoke/` alone was not protected at all. Protection by import
order is not protection.

THE FIX, in two parts.

1. The guard and `pg_schema` move to `tests/_store_isolation.py`, registered
   from `pyproject.toml` as `-p tests._store_isolation`. A plugin rather than
   `tests/conftest.py` for two reasons: that conftest is already over the
   per-file line cap and this has no business making it worse, and `-p` loads
   earlier and unconditionally rather than depending on which directory the
   run happens to collect.

2. `tests/smoke/conftest.py`'s `comms_env` now takes `pg_schema`, so the deny
   path gets a REAL but throwaway schema. Declared BEFORE `disk_tmp`
   deliberately: `pg_schema` connects while HOME is still real, and `comms_env`
   goes on to sandbox HOME, which is where libpq looks for its password file.

VERIFIED, and the second line is the one that matters:

    tests/smoke/test_node_comms_e2e_http.py   23 passed
    same file, immediately again               23 passed

Two consecutive green runs inside the cool-down window is the property that
was broken. Before the fix the second run failed; before the guard, it failed
while also editing production.

The intermediate state is worth recording: with the guard extended but before
`comms_env` took `pg_schema`, the file went to 15 failed / 7 errors — the
tests hitting the unreachable DSN. That is the guard working. Loud failure
against a port that refuses is the designed outcome; the previous silence was
the defect.
